refactor: mutation functions

+++ apps/evm/src/utilities/generateTransactionDeadline.ts
@@ -1,4 +1,4 @@
 import { TRANSACTION_TIMEOUT_S } from 'constants/transactionTimeout';
 
 export const generateTransactionDeadline = () =>
-  Math.trunc(new Date().getTime() / 1000) + TRANSACTION_TIMEOUT_S;
+  BigInt(Math.trunc(new Date().getTime() / 1000) + TRANSACTION_TIMEOUT_S);
